+++ controllers/genreController.js
@@ -14,7 +14,7 @@ const getAllGenres = async (req, res) => {
 const getGenreById = async (req, res) => {
     try {
         const id = req.params.id
-        if(!id) return errorHandler(res, 404, "ID not found")
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const genre = await genreModel.findByPk(id)
         if(!genre) return errorHandler(res, 404, `Genre with ${id} not found`) 
         res.status(200).json(genre)
@@ -39,15 +39,25 @@ const createGenre = async (req, res) => {
 
 const updateGenre = async (req, res) => {
     try {
-        
+        const id = req.params.id
+        if(!id) return errorHandler(res, 400, "Invalid ID")
+        const data = req.body
+        if(!data) return errorHandler(res, 400, "Invalid data")
+        await genreModel.update(data, {where: {id : id}})
+        const updatedGenre = await genreModel.findOne(data, {where: {id : id}})
+        if(!updatedGenre) return errorHandler(res, 404, "Genre not found")
+        res.status(200).json({
+            message: `Genre with ID: ${id} successfully updated.`, 
+            genre: updatedGenre}); 
     } catch (error) {
-        
-    }
+        errorHandler(res, 400, "Failed to update genre")
+    } 
 }
 
 const deleteGenre = async (req, res) => {
     try {
         const id = req.params.id
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const deletedGenre = genreModel.destroy({where: {id: id}})
         res.status(200).json({
             message: `Genre with ID: ${id} successfully deleted.`,
