+++ controllers/userController.js
@@ -14,7 +14,7 @@ const getAllUsers = async (req, res) => {
 const getUserById = async (req, res) => {
     try {
         const id = req.params.id    
-        if(!id) return errorHandler(res, 404, "ID not found")
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const user = await userModel.findByPk(id)
         if(!user) return errorHandler(res, 404, `User with ${id} not found`)
         res.status(200).json(user)
@@ -40,20 +40,15 @@ const createUser = async(req, res) => {
 const updateUser = async(req, res) => {
     try {
         const id = req.params.id
-        if(!id) return errorHandler(res, 404, "Invalid ID")
-            console.log(id);
-            
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const data = req.body
         if(!data) return errorHandler(res, 400, "Invalid data")
-            console.log(data);
-            
-        const user = await userModel.findByPk(id)
-        console.log(user);
-        
-        if(!user) return errorHandler(res, 404, "User not found")
-        await user.update(data)
-        if(!updatesUser) return errorHandler(res, 400, "Failed to update user")
-        res.status(200).json({message: `User with ID: ${id} successfully updated.`, user: updatesUser}); 
+        await userModel.update(data, {where: {id : id}})
+        const updatedUser = await userModel.findOne(data, {where: {id : id}})
+        if(!updatedUser) return errorHandler(res, 404, "User not found")
+        res.status(200).json({
+            message: `User with ID: ${id} successfully updated.`, 
+            user: updatedUser}); 
     } catch (error) {
         errorHandler(res, 400, "Failed to update user")
     } 
@@ -62,6 +57,7 @@ const updateUser = async(req, res) => {
 const deleteUser = async(req, res) => {
     try {
         const id = req.params.id
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const deletedUser = userModel.destroy({where: {id: id}})
         res.status(200).json({
             message: `User with ID: ${id} successfully deleted.`, 
@@ -79,7 +75,7 @@ export default { getAllUsers, getUserById, createUser, updateUser, deleteUser }
 // const getUserById = async (req, res) => {
 //     try {
 //         const id = req.params.id    
-//         if(!id) return errorHandler(res, 404, "ID not found")
+//         if(!id) return errorHandler(res, 404, "Invalid ID")
 //         const user = await userModel.findByPk(id)
 //         // const user = await userModel.findOne({where: { id: id }})
 //         if(!user) return errorHandler(res, 404, `User with ${id} not found`)
