+++ controllers/wishlistController.js
@@ -14,7 +14,7 @@ const getAllWishlists = async (req, res) => {
 const getWishlistById = async (req, res) => {
     try {
         const id = req.params.id
-        if(!id) return errorHandler(res, 404, "ID not found")
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const wishlist = wishlistModel.findByPk(id)
         if(!wishlist) return errorHandler(res, 404, `Wishlist with ${id} not found`)
         res.status(200).json(wishlist)
@@ -39,15 +39,25 @@ const createWishlist = async (req, res) => {
 
 const updateWishlist = async (req, res) => {
     try {
-        
+        const id = req.params.id
+        if(!id) return errorHandler(res, 400, "Invalid ID")
+        const data = req.body
+        if(!data) return errorHandler(res, 400, "Invalid data")
+        await wishlistModel.update(data, {where: {id : id}})
+        const updatedWishlist = await wishlistModel.findOne(data, {where: {id : id}})
+        if(!updatedWishlist) return errorHandler(res, 404, "Wishlist not found")
+        res.status(200).json({
+            message: `Wishlist with ID: ${id} successfully updated.`, 
+            book: updatedWishlist}); 
     } catch (error) {
-        
-    }
+        errorHandler(res, 400, "Failed to update wishlist")
+    } 
 }
 
 const deleteWishlist = async (req, res) => {
     try {
         const id = req.params.id
+        if(!id) return errorHandler(res, 400, "Invalid ID")
         const deletedWishlist = wishlistModel.destroy({where: {id: id}})
         res.status(200).json({
             message: `Wishlist with ID: ${id} successfully deleted.`,
